# components/manager/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Fetcher Manager component",
        "title": "Fetcher Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:4006",
    "basePath": "/",
    "paths": {
...
        "/v1/management/connections/unassigned": {
            "get": {
                "description": "List connections that have no product assigned, useful for migration purposes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Migration"
                ],
                "summary": "List unassigned connections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "X-Organization-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (minimum 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size (default 50, max 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Pagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.ConnectionResponse"
                                            }
                                        },
                                        "limit": {
                                            "type": "integer"
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/v1/management/connections/{id}/assign": {
            "post": {
                "description": "Associate an unassigned connection to a product. This is a one-time, irreversible operation for migration purposes. The product name must be provided via the X-Product-Name header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Migration"
                ],
                "summary": "Assign connection to product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "X-Organization-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Product name to assign",
                        "name": "X-Product-Name",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.ConnectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Pagination": {
            "description": "Pagination is the struct designed to store the pagination data of an entity list.",
            "type": "object",
            "properties": {
                "items": {},
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "github_com_LerianStudio_fetcher_pkg_model.ConnectionResponse": {
            "type": "object",
            "properties": {
                "configName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "databaseName": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "port": {
                    "type": "integer"
                },
                "productName": {
                    "type": "string"
                },
                "ssl": {
                    "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.SSLResponse"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
...
        "github_com_LerianStudio_fetcher_pkg_model.SSLResponse": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string"
                }
            }
        },
...
        "pkg.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                },
                "err": {},
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}